created RowsLayer cmp
