+++ src/components/Grid/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Row from './cmps/Row'
+import RowsLayer from './cmps/RowsLayer'
 
 const Grid = ({
     storyNodes,
@@ -11,28 +11,6 @@ const Grid = ({
     yCells,
 }) => {
 
-    const renderRows = () => {
-        const rowNums = [...Array(yCells).keys()].map(i => i + 1)
-        const rowCenter = Math.ceil(yCells / 2)
-        const cellWidth = width / xCells
-        const cellHeight = height / yCells
-        return rowNums.map(( rowNum ) => {
-            const rowGridY = ((rowNum - rowCenter) * -1) + gridViewCenter.y
-            const rowNodes = storyNodes.filter(node => node.grid_y === rowGridY)
-            return (
-                <Row
-                    key={rowNum}
-                    cellWidth={cellWidth}
-                    cellHeight={cellHeight}
-                    rowNum={rowNum}
-                    xCells={xCells}
-                    rowNodes={rowNodes}
-                    gridViewCenterX={gridViewCenter.x}
-                />
-            )
-        })
-    }
-
     return <svg
         version="1.1"
         baseProfile="full"
@@ -42,9 +20,14 @@ const Grid = ({
         viewBox={`0 0 ${width} ${height}`}
         preserveAspectRatio="xMinYMin meet"
     >
-        <g>
-            {renderRows()}
-        </g>
+        <RowsLayer
+            width={width}
+            height={height}
+            yCells={yCells}
+            xCells={xCells}
+            gridViewCenter={gridViewCenter}
+            storyNodes={storyNodes}
+        />
     </svg>
 }
